added logic for change in buy command along with reset deposit for user

+++ src/services/transaction/transaction.application/Features/Transactions/Commands/Deposit/AcceptedCent.cs
@@ -8,7 +8,7 @@ namespace transaction.application.Features.Transactions.Commands.Deposit;
 
 public static class AcceptedCent
 {
-    public static int[] cents = new int[] { 5, 10, 20, 50, 100 };
+    public static double[] cents = new double[] { 5, 10, 20, 50, 100 };
     public static bool CheckCent(int value)
     {
         if (cents.Any(c => c == value))
